fixed jira bug

+++ app/src/main/java/com/jetsynthesys/rightlife/ai_package/ui/eatright/fragment/DishLogEditFragment.kt
@@ -358,7 +358,7 @@ class DishLogEditFragment : BaseFragment<FragmentDishBinding>() {
         }
         if (!isEdit){
             if (snapRecipeData.mealQuantity != null ){
-                quantityEdit.setText(snapRecipeData.mealQuantity.toInt().toString())
+                quantityEdit.setText(snapRecipeData.mealQuantity.toString())
             }
         }
         var imageUrl : String? = ""
